@@ -1,34 +1,33 @@
 ﻿using DesignPatterns.Creational.AbstractFactory.Abstractions;
 
-namespace DesignPatterns.Creational.AbstractFactory.Models
+namespace DesignPatterns.Creational.AbstractFactory.Models;
+
+/// <inheritdoc />
+public class ModernChair : IChair
 {
     /// <inheritdoc />
-    public class ModernChair : IChair
-    {
-        /// <inheritdoc />
-        public FurnitureStyle GetFurnitureStyle() => FurnitureStyle.Modern;
-
-        /// <inheritdoc />
-        public FurnitureType GetFurnitureType() => FurnitureType.Chair;
-    }
+    public FurnitureStyle GetFurnitureStyle() => FurnitureStyle.Modern;
 
     /// <inheritdoc />
-    public class ModernCoffeeTable : ICoffeeTable
-    {
-        /// <inheritdoc />
-        public FurnitureStyle GetFurnitureStyle() => FurnitureStyle.Modern;
+    public FurnitureType GetFurnitureType() => FurnitureType.Chair;
+}
 
-        /// <inheritdoc />
-        public FurnitureType GetFurnitureType() => FurnitureType.CoffeeTable;
-    }
+/// <inheritdoc />
+public class ModernCoffeeTable : ICoffeeTable
+{
+    /// <inheritdoc />
+    public FurnitureStyle GetFurnitureStyle() => FurnitureStyle.Modern;
 
     /// <inheritdoc />
-    public class ModernSofa : ISofa
-    {
-        /// <inheritdoc />
-        public FurnitureStyle GetFurnitureStyle() => FurnitureStyle.Modern;
+    public FurnitureType GetFurnitureType() => FurnitureType.CoffeeTable;
+}
 
-        /// <inheritdoc />
-        public FurnitureType GetFurnitureType() => FurnitureType.Sofa;
-    }
+/// <inheritdoc />
+public class ModernSofa : ISofa
+{
+    /// <inheritdoc />
+    public FurnitureStyle GetFurnitureStyle() => FurnitureStyle.Modern;
+
+    /// <inheritdoc />
+    public FurnitureType GetFurnitureType() => FurnitureType.Sofa;
 }
