@@ -1,18 +1,16 @@
-﻿
-using DesignPatterns.Creational.Builder.Models;
+﻿using DesignPatterns.Creational.Builder.Models;
 
-namespace DesignPatterns.Creational.Builder.Builders
+namespace DesignPatterns.Creational.Builder.Builders;
+
+/// <inheritdoc />
+public class PizzaHawaiRecipe : PizzaRecipe
 {
     /// <inheritdoc />
-    public class PizzaHawaiRecipe : AbstractPizzaRecipe
-    {
-        /// <inheritdoc />
-        public PizzaHawaiRecipe() : base("Hawai") { }
+    public PizzaHawaiRecipe() : base("Hawai") { }
 
-        /// <inheritdoc />
-        public override void AddSauce() => Pizza.Sauce = Sauce.Tomato;
+    /// <inheritdoc />
+    public override void AddSauce() => Pizza.Sauce = Sauce.Tomato;
 
-        /// <inheritdoc />
-        public override void AddIngredients() => Pizza.Ingredients = new[] {Ingredient.Mozzarella, Ingredient.Ham, Ingredient.Pineapple};
-    }
+    /// <inheritdoc />
+    public override void AddIngredients() => Pizza.Ingredients = new[] {Ingredient.Mozzarella, Ingredient.Ham, Ingredient.Pineapple};
 }
