@@ -1,17 +1,16 @@
 ﻿using DesignPatterns.Creational.Builder.Models;
 
-namespace DesignPatterns.Creational.Builder.Builders
+namespace DesignPatterns.Creational.Builder.Builders;
+
+/// <inheritdoc />
+public class PizzaTonnoRecipe : PizzaRecipe
 {
     /// <inheritdoc />
-    public class PizzaTonnoRecipe : AbstractPizzaRecipe
-    {
-        /// <inheritdoc />
-        public PizzaTonnoRecipe() : base("Tonno") { }
+    public PizzaTonnoRecipe() : base("Tonno") { }
 
-        /// <inheritdoc />
-        public override void AddSauce() => Pizza.Sauce = Sauce.Tomato;
+    /// <inheritdoc />
+    public override void AddSauce() => Pizza.Sauce = Sauce.Tomato;
 
-        /// <inheritdoc />
-        public override void AddIngredients() => Pizza.Ingredients = new[] { Ingredient.Mozzarella, Ingredient.Tuna, Ingredient.Union };
-    }
+    /// <inheritdoc />
+    public override void AddIngredients() => Pizza.Ingredients = new[] { Ingredient.Mozzarella, Ingredient.Tuna, Ingredient.Union };
 }
