@@ -1,32 +1,38 @@
 ﻿using DesignPatterns.Creational.Builder.Builders;
 using DesignPatterns.Creational.Builder.Models;
 
-namespace DesignPatterns.Creational.Builder
+namespace DesignPatterns.Creational.Builder;
+
+/// <summary>
+/// Chef who creates the pizza a.k.a. the director.
+/// </summary>
+public class Chef
 {
+    private PizzaRecipe _pizzaRecipe;
+
+    private Chef() { }
+
     /// <summary>
-    /// Chef who creates the pizza a.k.a. the director.
+    /// Get the chef of this establishment.
     /// </summary>
-    public class Chef
+    public static Chef Instance { get; } = new();
+
+    /// <summary>
+    /// Give the pizza recipe to the chef.
+    /// </summary>
+    /// <param name="pizzaRecipe"><see cref="PizzaRecipe"/></param>
+    public void GiveRecipe(PizzaRecipe pizzaRecipe) => _pizzaRecipe = pizzaRecipe;
+
+    /// <summary>
+    /// Let's bake the pizza based on the last given recipe.
+    /// </summary>
+    /// <returns><see cref="Pizza"/></returns>
+    public Pizza BakePizza()
     {
-        private AbstractPizzaRecipe _pizzaRecipe;
+        _pizzaRecipe.LayoutDough();
+        _pizzaRecipe.AddSauce();
+        _pizzaRecipe.AddIngredients();
 
-        /// <summary>
-        /// Give the pizza recipe to the chef.
-        /// </summary>
-        /// <param name="pizzaRecipe"><see cref="AbstractPizzaRecipe"/></param>
-        public void GiveRecipe(AbstractPizzaRecipe pizzaRecipe) => _pizzaRecipe = pizzaRecipe;
-
-        /// <summary>
-        /// Let's bake the pizza based on a given recipe.
-        /// </summary>
-        /// <returns><see cref="Pizza"/></returns>
-        public Pizza BakePizza()
-        {
-            _pizzaRecipe.LayoutDough();
-            _pizzaRecipe.AddSauce();
-            _pizzaRecipe.AddIngredients();
-
-            return _pizzaRecipe.BakePizza();
-        }
+        return _pizzaRecipe.BakePizza();
     }
 }
