@@ -1,14 +1,13 @@
-﻿namespace DesignPatterns.Creational.FactoryMethod.Abstractions
+﻿namespace DesignPatterns.Creational.FactoryMethod.Abstractions;
+
+/// <summary>
+/// Connector that can handle transportations of packages
+/// </summary>
+public interface IConnector
 {
     /// <summary>
-    /// Connector that can handle transportations of packages
+    /// Create a transport instance.
     /// </summary>
-    public interface IConnector
-    {
-        /// <summary>
-        /// Create a transport instance.
-        /// </summary>
-        /// <returns><see cref="ITransport"/></returns>
-        ITransport CreateTransport();
-    }
+    /// <returns><see cref="ITransport"/></returns>
+    ITransport CreateTransport();
 }
