@@ -1,13 +1,12 @@
-﻿namespace DesignPatterns.Creational.FactoryMethod.Abstractions
+﻿namespace DesignPatterns.Creational.FactoryMethod.Abstractions;
+
+/// <summary>
+/// Method of transportation.
+/// </summary>
+public interface ITransport
 {
     /// <summary>
-    /// Method of transportation.
+    /// Send a package.
     /// </summary>
-    public interface ITransport
-    {
-        /// <summary>
-        /// Send a package.
-        /// </summary>
-        void SendPackage();
-    }
+    IEnumerable<string> SendPackage();
 }
