@@ -1,12 +1,11 @@
 ﻿using DesignPatterns.Creational.FactoryMethod.Abstractions;
 using DesignPatterns.Creational.FactoryMethod.Models;
 
-namespace DesignPatterns.Creational.FactoryMethod.Connectors
+namespace DesignPatterns.Creational.FactoryMethod.Connectors;
+
+/// <inheritdoc />
+public class TcpConnector : IConnector
 {
     /// <inheritdoc />
-    public class TcpConnector : IConnector
-    {
-        /// <inheritdoc />
-        public ITransport CreateTransport() => new Tcp();
-    }
+    public ITransport CreateTransport() => new Tcp();
 }
