@@ -1,12 +1,11 @@
 ﻿using DesignPatterns.Creational.FactoryMethod.Abstractions;
 using DesignPatterns.Creational.FactoryMethod.Models;
 
-namespace DesignPatterns.Creational.FactoryMethod.Connectors
+namespace DesignPatterns.Creational.FactoryMethod.Connectors;
+
+/// <inheritdoc />
+public class UdpConnector : IConnector
 {
     /// <inheritdoc />
-    public class UdpConnector : IConnector
-    {
-        /// <inheritdoc />
-        public ITransport CreateTransport() => new Udp();
-    }
+    public ITransport CreateTransport() => new Udp();
 }
