@@ -1,20 +1,15 @@
-﻿using System;
-using System.Collections.Generic;
-using DesignPatterns.Creational.FactoryMethod.Abstractions;
+﻿using DesignPatterns.Creational.FactoryMethod.Abstractions;
 
-namespace DesignPatterns.Creational.FactoryMethod.Models
+namespace DesignPatterns.Creational.FactoryMethod.Models;
+
+/// <inheritdoc />
+public class Tcp : ITransport
 {
     /// <inheritdoc />
-    public class Tcp : ITransport
+    public IEnumerable<string> SendPackage()
     {
-        /// <inheritdoc />
-        public void SendPackage() =>
-            new List<string>
-                {
-                    $"{nameof(Tcp).ToUpper()} - Send SYN",
-                    $"{nameof(Tcp).ToUpper()} - SYN received",
-                    $"{nameof(Tcp).ToUpper()} - Send ACK"
-                }
-                .ForEach(Console.WriteLine);
+        yield return $"{nameof(Tcp).ToUpper()} - Send SYN";
+        yield return $"{nameof(Tcp).ToUpper()} - SYN received";
+        yield return $"{nameof(Tcp).ToUpper()} - Send ACK";
     }
 }
