@@ -1,37 +1,34 @@
-﻿using System;
-using DesignPatterns.Structural.Adapter.Shapes;
+﻿using DesignPatterns.Structural.Adapter.Shapes;
 
-namespace DesignPatterns.Structural.Adapter.Holes
+namespace DesignPatterns.Structural.Adapter.Holes;
+
+/// <summary>
+/// Details of a round hole.
+/// </summary>
+public class RoundHole
 {
     /// <summary>
-    /// Details of a round hole.
+    /// Default constructor
     /// </summary>
-    public class RoundHole
+    /// <param name="radius"><see cref="Radius"/></param>
+    public RoundHole(int radius)
     {
-        /// <summary>
-        /// Default constructor
-        /// </summary>
-        /// <param name="radius"><see cref="Radius"/></param>
-        public RoundHole(int radius)
-        {
-            Radius = radius;
-        }
+        Radius = radius;
+    }
 
-        /// <summary>
-        /// Radius of the hole.
-        /// </summary>
-        public int Radius { get; set; }
+    /// <summary>
+    /// Radius of the hole.
+    /// </summary>
+    public int Radius { get; set; }
 
-        /// <summary>
-        /// 
-        /// </summary>
-        /// <param name="peg"></param>
-        /// <returns></returns>
-        public void Fits(RoundPeg peg)
-        {
-            Console.WriteLine($"Peg {peg}.");
-            Console.WriteLine($"But does it fit in hole with radius {Radius}? {(Radius >= peg.Radius ? "Yes!" : "No!")}");
-            Console.WriteLine();
-        }
+    /// <summary>
+    /// Whether the peg will fit in the round hole or not.
+    /// </summary>
+    /// <param name="peg"><see cref="RoundPeg"/></param>
+    public void Fits(RoundPeg peg)
+    {
+        Console.WriteLine($"Peg {peg}.");
+        Console.WriteLine($"But does it fit in hole with radius {Radius}? {(Radius >= peg.Radius ? "Yes!" : "No!")}");
+        Console.WriteLine();
     }
 }
