@@ -1,28 +1,16 @@
-﻿using System;
+﻿using DesignPatterns.Creational.Builder;
 using DesignPatterns.Creational.Builder.Builders;
 
-namespace DesignPatterns.Creational.Builder
+// Ask for several pizza's.
+GetPizza(new PizzaHawaiRecipe());
+GetPizza(new PizzaMargheritaRecipe());
+GetPizza(new PizzaPepperoniRecipe());
+GetPizza(new PizzaProsciuttoFunghiRecipe());
+GetPizza(new PizzaTonnoRecipe());
+
+static void GetPizza(PizzaRecipe pizzaRecipe)
 {
-    internal static class Program
-    {
-        private static void Main()
-        {
-            // Hire a new chef
-            var chef = new Chef();
+    Chef.Instance.GiveRecipe(pizzaRecipe);
 
-            // Ask for several pizza's.
-            chef.GetPizza(new PizzaHawaiRecipe());
-            chef.GetPizza(new PizzaMargheritaRecipe());
-            chef.GetPizza(new PizzaPepperoniRecipe());
-            chef.GetPizza(new PizzaProsciuttoFunghiRecipe());
-            chef.GetPizza(new PizzaTonnoRecipe());
-        }
-
-        private static void GetPizza(this Chef chef, AbstractPizzaRecipe pizzaRecipe)
-        {
-            chef.GiveRecipe(pizzaRecipe);
-
-            Console.WriteLine(chef.BakePizza());
-        }
-    }
+    Console.WriteLine(Chef.Instance.BakePizza());
 }
