@@ -1,16 +1,7 @@
-﻿using System;
+﻿using DesignPatterns.Creational.Singleton;
 
-namespace DesignPatterns.Creational.Singleton
-{
-    internal class Program
-    {
-        private static void Main()
-        {
-            Console.WriteLine(Sun.Instance);
-            Console.WriteLine(Sun.Instance);
-            Console.WriteLine(Sun.Instance);
-            Console.WriteLine(Sun.Instance);
-            Console.WriteLine(Sun.Instance);
-        }
-    }
-}
+Console.WriteLine(Sun.Instance);
+Console.WriteLine(Sun.Instance);
+Console.WriteLine(Sun.Instance);
+Console.WriteLine(Sun.Instance);
+Console.WriteLine(Sun.Instance);
